Offer user option to move files to a directory instead of deleting

+++ main.go
@@ -8,15 +8,30 @@ import (
 )
 
 func main() {
-	var path string
+	var path, search, action string
+
 	fmt.Println("Enter the path to the folder you want to search and delete files in:")
 	fmt.Scan(&path)
 
-	var search string
 	fmt.Println("Enter the string that will be used to search and delete files:")
 	fmt.Scan(&search)
 
-	fmt.Printf("\nPath: %s\nSearch: %s\n\n", path, search)
+	fmt.Println("Permanently delete files? (y/n)\nSelecting \"no\" will move files to a new folder.")
+	fmt.Scan(&action)
+
+	fmt.Printf("\nPath: %s\nSearch: %s\nAction: %s\n\n", path, search, action)
+
+	deleteDir := path + "/_deleted"
+
+	// If user doesn't want to delete files, create a new folder to move them to
+	if action == "n" {
+		err := os.MkdirAll(deleteDir, os.ModePerm)
+		if err != nil {
+			fmt.Println(err)
+		} else {
+			fmt.Println("Directory created:", deleteDir)
+		}
+	}
 
 	// List all files under the path recursively, skipping folders
 	err := filepath.Walk(path, func(path string, info os.FileInfo, err error) error {
@@ -29,20 +44,29 @@ func main() {
 			return nil
 		}
 
-		// Check if the file matches the search string
+		// Check if file matches the search string
 		if strings.Contains(info.Name(), search) {
 			fmt.Printf("File: %q\n", info.Name())
 
-			// Delete the file
-			err := os.Remove(path)
-			if err != nil {
-				fmt.Println(err)
+			if action == "y" {
+				// Delete file
+				err := os.Remove(path)
+				if err != nil {
+					fmt.Println(err)
+				}
+				// TODO: Log file names
+
+			} else {
+				// Move file to a new folder
+				err = os.Rename(path, deleteDir+"/"+info.Name())
+				if err != nil {
+					fmt.Println(err)
+				}
 			}
 
-			// TODO: Rather than permanently deleting the file, move it to a new folder called "deleted"
-			// TODO: This should actually be an option; ask user if they want to permanently delete or not
-			// TODO: If deleting, log the file names
 			return nil
+		} else {
+			fmt.Println("No match found")
 		}
 
 		return nil
